+++ CHANGELOG.md
@@ -9,13 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Connection.authenticate_oidc()`: add argument to set maximum device code flow poll time
+- Show progress bar while waiting for OIDC authentication with device code flow,
+  including special mode for in Jupyter notebooks.
+  ([#237](https://github.com/Open-EO/openeo-python-client/issues/237))
+- Basic support for `load_stac` process with `Connection.load_stac()`
+  ([#425](https://github.com/Open-EO/openeo-python-client/issues/))
+
 ### Changed
 
 ### Removed
 
 ### Fixed
 
-
+- Include "scope" parameter in OIDC token request with client credentials grant.
+- Support fractional seconds in `Rfc3339.parse_datetime`
+  ([#418](https://github.com/Open-EO/openeo-python-client/issues/418))
 
 ## [0.16.0] - 2023-04-17 - "SRR5" release